Hardcodes filter dialog, basic functionality intact
