+++ packages/filebrowser-extension/schema/widget.json
@@ -9,8 +9,7 @@
         "rank": 10
       },
       { "name": "uploader", "rank": 20 },
-      { "name": "refresh", "command": "filebrowser:refresh", "rank": 30 },
-      { "name": "fileNameSearcher", "rank": 40 }
+      { "name": "refresh", "command": "filebrowser:refresh", "rank": 30 }
     ]
   },
   "jupyter.lab.menus": {
